+++ CHANGELOG.md
@@ -47,6 +47,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Incorrect altered state of masters after loading in some cases.
 - Setting LoadingState.Loaded to DataObject after loading.
 - Objects updating order if exists cycle in dependencies graph of them.
+- DbTransactionWrapper commit and rollback over expired connections.
+- Null GetHandler or SetHandler via cache dictionary. 
+- Rethrowing exception while handling special scenario via UpdateObjects.
+- Fix loading __PrimaryKey property of NotStored master.
 
 ### Security